# tanks_backend/src/main/resources/documentation/openAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Tanks Online API",
    "description": "API documentation for ATA systems multiplayer game",
    "contact": {
      "name": "Antitalent Team",
      "email": "omazicantonio7@gmail.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/user/{userName}/status": {
      "post": {
        "tags": [
          "User Management"
        ],
        "summary": "Update user status",
        "description": "Updates a user's online status by username",
        "operationId": "updateUserStatus",
        "parameters": [
          {
            "name": "userName",
            "in": "path",
            "description": "Username of the user to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "newStatus",
            "in": "query",
            "description": "New status to set",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "ONLINE",
                "OFFLINE",
                "IN_GAME",
                "BANNED",
                "INACTIVE"
              ]
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status update request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "200": {
            "description": "Status updated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GrantedAuthority": {
        "type": "object",
        "properties": {
          "authority": {
            "type": "string"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "passwordHash": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "GUEST",
              "PLAYER",
              "ADMIN"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ONLINE",
              "OFFLINE",
              "IN_GAME",
              "BANNED",
              "INACTIVE"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastLogin": {
            "type": "string",
            "format": "date-time"
          },
          "userStats": {
            "$ref": "#/components/schemas/UserStats"
          },
          "password": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "authorities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GrantedAuthority"
            }
          },
          "accountNonExpired": {
            "type": "boolean"
          },
          "accountNonLocked": {
            "type": "boolean"
          },
          "credentialsNonExpired": {
            "type": "boolean"
          }
        }
      },
      "UserStats": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "wins": {
            "type": "integer",
            "format": "int32"
          },
          "losses": {
            "type": "integer",
            "format": "int32"
          },
          "kills": {
            "type": "integer",
            "format": "int32"
          },
          "deaths": {
            "type": "integer",
            "format": "int32"
          },
          "matchesPlayed": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}